moved config.js to root
