+++ src/ActionModal/utils/ActionManager.js
@@ -1,5 +1,5 @@
 import Cosmos from "@lunie/cosmos-api"
-import config from "src/config"
+import config from "src/../config"
 import { getSigner } from "./signer"
 import transaction from "./transactionTypes"
 import { uatoms } from "scripts/num"
