+++ src/gql/apollo.js
@@ -2,7 +2,7 @@ import Vue from "vue"
 import ApolloClient from "apollo-boost"
 import { InMemoryCache } from "apollo-cache-inmemory"
 import VueApollo from "vue-apollo"
-import config from "src/config"
+import config from "src/../config"
 
 Vue.use(VueApollo)
