+++ src/vuex/modules/ledger.js
@@ -1,4 +1,4 @@
-import config from "src/config"
+import config from "src/../config"
 
 export default () => {
   const emptyState = {}
